+++ src/js/utils/utils.ts
@@ -1,3 +1,4 @@
+import { useEffect, useRef } from 'react';
 import { Config } from '~/src/js/config';
 
 export const fetchConfig = async (PLUGIN_ID: string): Promise<Config | null> => {
@@ -6,4 +7,40 @@ export const fetchConfig = async (PLUGIN_ID: string): Promise<Config | null> => 
   if (!rawConfig) return null;
   const config = JSON.parse(rawConfig);
   return config;
+};
+
+export const useRecursiveTimeout = <T extends unknown>(
+  callback: () => Promise<T> | (() => void),
+  delay: number | null,
+): void => {
+  const savedCallback = useRef(callback);
+
+  // Remember the latest callback.
+  useEffect(() => {
+    savedCallback.current = callback;
+  }, [callback]);
+
+  // Set up the timeout loop.
+  useEffect(() => {
+    let id: NodeJS.Timeout;
+    function tick() {
+      const ret = savedCallback.current();
+
+      if (ret instanceof Promise) {
+        ret.then(() => {
+          if (delay !== null) {
+            id = setTimeout(tick, delay);
+          }
+        });
+      } else {
+        if (delay !== null) {
+          id = setTimeout(tick, delay);
+        }
+      }
+    }
+    if (delay !== null) {
+      id = setTimeout(tick, delay);
+      return () => id && clearTimeout(id);
+    }
+  }, [delay]);
 };
