+++ src/types/kintone.d.ts
@@ -1,0 +1,3 @@
+declare namespace kintone {
+  const $PLUGIN_ID: string;
+}
